+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added very basic `WeakMap` polyfill
 * Add `Function.getArgumentNames`
 * Allow ability to add Function driers
+* Added `Number.isNumeric(input)`
 
 ## 0.3.7 (2017-07-03)